Updated Game Structure
Created a games directory and landing page so that the game does not have to be loaded on the projects page.

The games page was left out of the navbar as they should be available on the projects page for the foreseeable future, but it can be added if it becomes a central feature of the site.