YouTube Embed Update 3 (Final)
Created a custom class and script to "lazy-load" the iframes in a more manual way. The script attaches an eventListener to scrolling and only loades those iframes which are visible after being scrolled upon. This might not be great for scale due to never deloading the iframes, but it definitely cuts down on initial loading times.